👌 IMPROVE: a11y
I just submitted a talk and the message was not clear enough to help me understand if the talk was submitted. So, I added `SUBMITTED` to the mix. I think the message should also have a green background but that's a story for another PR.

Thank you for creating this awesome resource.